rename requirements in store to weaponRequirements
