+++ src/data/weaponRequirements.js
